install: move exported BLS/karg functions into io
Several BLS helpers have grown into generic library functions, so they
really shouldn't stay in install.
